First draft of image tabstrip
(Relates to #93)

Visual appearance is very rough, basically the bare minimum necessary to
test hover/active/inactive, etc.  That said, I may keep this approach as
a fallback for unthemed environments... not sure I want to put much work
into that fringe case.

Also included in this commit are myriad small fixes for other window
issues, as seems to be the pattern lately.  I've also added support for
alpha premultiplication to the pdLayer class, which was needed to make
my life easier when drawing transparent thumbnails onto the image
taskbar.