For floating-images + docked-toolbars, enforce MDI-like behavior
Specifically, restrict image windows to the main form's client area.

This solves a number of layering issues relating to docked toolbars (and
associated windows, like toolbar tooltips) when images are floating.
While I'm not a fan of the MDI paradigm, it's hard to argue with the
results of this commit: it not only solves a number of tricky bugs, but
also produces a familiar paradigm for users of old PD versions.  I think
that's a fair trade.

This commit also fixes some troubling redraw issues on non-Aero themes,
and I *think* Win 7 basic and all classic/high-contrast/etc variations
are now rendering correctly, regardless of float/docked state of both
toolbars and image windows.