Begin preferences purge
(This is going to be an ongoing project.)

When I first created a full-featured preferences handler, it was
tempting to dump a whole bunch of options there just because I could.
Now that I'm older and wiser (?), I realize that I have frequently used
preferences as a way to avoid making hard design decisions.

Slowly but surely, I hope to remove a number of superfluous preferences
from the program.  This commit removes a few dumb interface preferences
that should always be enabled (remember window location and dynamic
taskbar icons).