Undo/Redo engine overhaul
(I have wanted to do this for a very, very long time.)

PD has had Undo/Redo support since pretty much its inception.  The first
incarnation stored a byte array of the prior image state in memory,
which allowed for a single undo operation to be performed.  Over time,
the Undo system has grown much more complex, but it has done so by
strapping ever more features onto that original flimsy framework.

Fast-forward to 2013, and PD's Undo/Redo code was a hot mess, with
functionality split across various modules and classes, with a bunch of
tangled functions repeating the same behaviors at each step of the
process, but not in a predictable way.

Thus the need for a rewrite.  As of this commit, all Undo/Redo data is
now encapsulated within a single pdUndo class.  Each pdImage object
contains one instance of pdUndo.  ALL Undo/Redo functionality is handled
through this class, specifically via something like
pdImages(imageIndex).undoManager.whatever.

This should make it significantly easier to debug Undo/Redo stuff in the
future, as well as expand its capabilities to include things like a full
history browser.

Some interface code remains in the pdUndo class for now.  I will be
removing this shortly, but it requires the creation of a new (and
much-improved) function for synchronizing the interface to the image's
current state.  Stay tuned.