Check for thunks before removing from subclasser collection
A "remove thunk from collection" line was throwing errors at me, so I
implemented a basic check to see if a thunk exists before removing it
from a collection.  That seemed to solve the problem.

Chances are that I'm attempting to unsubclass something I never
subclassed in the first place - will investigate, but as I see no harm
in this fix (and in fact, it makes the class more robust), I think I'll
leave it.