Provide unsaved changes icon on image tabstrip
I'm not 100% convinced that an asterisk is the best icon here, but as
many programs use an asterisk in the window caption for unsaved changes,
perhaps there is a natural association...?

I've tried using a disk icon, but it's too easily lost on dark images,
and it's confusing to use a disk icon to indicated "not saved".
Similarly, using a disk icon to indicate "has been saved" is
non-standard and just as confusing!

Will stick with the asterisk for now, but I remain on the lookout for
something better...