Overhauled (and much-improved) JPEG export dialog
New features include:

1) Live previews!  I have wanted this *for so long*.  All work is done
in memory, so the hard drive is not touched by the previewer.  Though a
small addition, I consider this one of the best things I have added to
PD in a long time in terms of my personal workflow.

2) Command bar added, so full support for custom presets, auto-save
last-used settings, etc.

3) Necessary interface update to account for the new features.  I'm not
thrilled about the odd aspect size of the new dialog, but the preview is
worthless if it's too small, so this will have to do for now.