Rewritten Recent Files manager, including new preference for max recent file count
Every release I like to revisit some old chunks of code, not necessarily
because they are broken, but because it's nice to bring a fresh set of
eyes to something I haven't investigated for awhile.  This week, that
honor fell to the Recent Files module.

The module has been rewritten as a class, and most functions have been
rewritten and recommented for clarity and style improvements.  MRU data
has also been moved out of the core preferences file, and into a
separate XML file saved in the /Presets folder.  This is much better
from an organizational standpoint, as I can now wipe the file at will
without affecting other preferences.

Speaking of preferences, I've added a new one for changing the max
number of files saved to the recent files list; this can be anywhere
from 1 to 32 (32 makes more sense on XP, where icons are only 16x16),
and the preference can be changed without requiring a reboot.

I've considered an option for 0, which effectively disallows saving of
files to the Recent Files menu, which could be a useful privacy setting
- however, that breaks the MRU code at present so I'll need to rewrite a
few things to allow for a max value of 0.