Greatly improve GDI+ image load function
The original version of this function was the first bit of GDI+ code I
ever wrote, and it was not pretty.  (Images were passed back to the DIB
via StdPicture!  Ouch!)  The new codepath uses a year's worth of
accumulated GDI+ knowledge to greatly improve performance for both 24
and 32bpp images, including relying on GDI+ itself to unpremultiply
alpha as necessary.  (In the next few weeks, I will toggle this to apply
premultiplication, which is part of why I wanted to rewrite the code in
advance!)

This was also a good opportunity for me to test LockBits, which works
very well for copying large amounts of raw bitmap data between GDI+
objects and DIBs.