Mark all output files as sRGB color space
One line of code to implement this.  I *really* love ExifTool.

What makes this doubly lovely is that ExifTool automatically figures out
if an image format supports a color-space type tag, then writes it in
whichever format is appropriate to that file.  We don't have to deal
with JPEG/PNG/TIFF differences at all.