Minor bugfixes to color management (and more)
Included is a fix for mousewheel event handling.  Under certain
windowing circumstances, it is possible for the main window to have
focus (but not to appear that way), which prevents the child window from
receiving mousewheel events.  The main window now receives these events
and passes them along to a child window if a child window is present.