Add individual text boxes and picture boxes to color selector
Works wonderfully!

I really do like the GIMP color selector, so I am rather shamelessly
ripping it off in its entirety.  One of the things I like most about it
is that it clearly shows the impact of changing any one of RGB or HSV,
which makes it very simple to acquire colors "like" the one you
currently have, but differing in some singular way.

Next up is RGB <-> Hex, which shouldn't be too tricky... "in theory".