Assign hand cursor to clickable picture boxes on color selector
This wraps up the whirlwind creation of a custom color selector.  It's
in "workable" state for now, which means I can leave it for a bit to
focus on some other tasks.  (There's still a lot of details left to
tweak/add, but I mostly wanted to implement it to test out color
management.)

All-in-all, not too shabby for an evening's worth of work.