Add file/path tooltips to the image thumbnail bar
Been wanting to do this for awhile, but the poor tooltip class I'm using
is a debacle.  I've rewritten so much of it that I should probably label
it my own work...ugh.  Unbeknownst to me, the original creator decided
to implement tooltip titles as a custom string that he simply prepends
to all tooltips.  I have rewritten this so that titles actually set the
tooltip title text property properly.  (Say that 10x fast!)