Temporary Print dialog solution
So, building a custom Print dialog is a *huge* undertaking.  Trying to
provide a powerful print experience requires a ton of obscure API calls
that do not play nicely in VB, and while I think it's possible to do,
it's going to be a lot of work.

Because I have higher priorities at the moment, I am instituting a
temporary Print dialog solution for users of Vista+.  PhotoDemon's Print
dialog now re-routes through the stock Windows Photo Print dialog, which
has its problems, but is significantly better than the current PD print
dialog.

Unfortunately, XP users are stuck with the old dialog at present,
because XP does not have an easily accessible "print" verb for image
files, despite having a context menu entry for it.

I will keep pecking away at the new print dialog, but not as a top
priority.