Use pre-multiplied alpha by default
(Relates to #96)

This may not seem like much of a change, but it's actually a ground-up
rewrite of the way PD handles 32bpp data.  Premultiplied alpha is
required for hardware-accelerated transparent rendering (via
AlphaBlend), and with layers support coming, it was impractical to
continue doing my own manual compositing.

This commit primarily affects the handling of premultiplied data at
load-time.  Much more work is required to fix the additional
implications throughout the software.