Rewrite alpha checkerboard rendering system
Previously, temporary copies of 32bpp images self-composited against a
checkerboard background on-demand (typically just before rendering to
the screen).  This required the use of custom pre-multiplication code
that had to be re-applied every time an on-screen copy of a 32bpp image
was updated.

Now, we simply cache a 2x2 grid of the current alpha pattern, and render
it on-demand using PatBlt, then overlay the 32bpp data using AlphaBlend.
This provides much better (faster) results.