Rewrite image tool pre- and post-processors to work with new alpha code
(Relates to #96)

For the record, this commit also includes a number of high-level changes
to the pre/post-processors to make the eventual addition of pan/zoom
previewing possible.

While pre-multiplied alpha is great for improving performance of
on-screen rendering, it must be removed prior to editing image data.
Otherwise, the RGB data will be incorrect and simple actions like
"brighten" will return faulty data (because the RGB data includes alpha
data from the premultiplication).

So for 32bpp data, the pre-processor now removes premultiplication,
while the post-processor adds it back in.

I have also completely rewritten the way selected regions are handled.
When a selection is active, the full selected rect - without
modification - is sent to the image tool, even during previewing.  When
the tool completes, the post-processor now performs a full composite of
the selected area, then renders that to the preview area.

Previously, only the selected area was displayed, but now, the user can
see how the selected area will composite against the final image, which
I think is much more helpful, particularly when working with feathered
selections.

A number of changes were also made to the fxPreview user control to
handle this new data correctly.