New Hardware Acceleration report in Tools -> Options -> Advanced
This is the kind of superfluous garbage I usually try to avoid, but the
Advanced options panel looked so bare, and I have some genuine curiosity
about hardware support for various GDI operations, especially as I look
toward implementing drawing and text tools.

GetDeviceCaps is used to probe the primary display device for various
support functions.  When I finally get around to rewriting PD's print
dialog, I'd like to extend this to accept arbitrary DCs, so we can also
probe the printer.

The current report format is painfully ugly, but I wanted it to be
copy-and-pastable so I can request it from a user if necessary.  I may
get around to "pretty-ifyng" this in the future, but it's not exactly a
high priority...