Solve flickering-on-resize issues
Relates to #98.

I am starting to believe that VB natively implements a bunch of things
that should be handled via the API.

As an example, setting ClipControls on the main form causes all kinds of
trouble with child windows, while not reducing flicker in the least.  I
previously assumed that property just set a matching GW_STYLE flag of
WS_CLIPCHILDREN - but that is not the case!  Manually setting a
WS_CLIPCHILDREN style has fixed all my flicker issues, so VB must do
something else entirely.  Good to know...?