Implement click-drag scrolling for the image tabstrip
Now I can finally get rid of that hideous horizontal scrollbar.  :)
Actually, it still exists, as it provides an easy way to track and allow
scrolling - but it is kept invisible.

This is also a nice bonus for touchscreen users!

The one downside to this approach is that it's not immediately obvious
that the list can be click-dragged... but I'm not sure how to avoid
that.