New window manager framework for changing image tabstrip alignment
As the tabstrip's location and size affects rendering of docked image
windows, this was a fairly comprehensive update.

This commit does not yet include an interface mechanism for changing the
tabstrip's alignment, nor does it include any code within the tabstrip
for drawing the image list in a vertical layout.  Those are coming
later.