Add visibility options for image tabstrip
I debated the merit of displaying these options to the user, and
eventually decided it was worth it.  Other toolbars can be selectively
displayed/hidden so it seems unnatural to omit the image tabstrip
(especially for those who seek a truly minimal interface).  Also, I'm
one of those weirdos who likes tabs always visible in my web browser,
even when only a single page is open, so I should extend the same option
to PD users.

Per PD convention, these options are exposed directly in the Window
menu, not in the Options dialog.