Implement all possible alignments for image tabstrip
Big project, glad to have it done.

This commit also includes reactivation of tabstrip mousewheel handling,
FYI.

There's still a bit of buggy mouse behavior to work out, particularly
when tracking tabstrip resizing as the tabstrip gets confused if the
user is resizing or scrolling.  I'll fix that in a bit.