Fix initial position setting of image tabstrip on fresh install
Dumb mistake on my part - apologies to anyone who's downloaded a fresh
copy of PD in the past few days, as the image tabstrip will be
completely botched.  The "set default position" code wasn't firing on
new installs, so the tabstrip failed to ever be aligned properly,
meaning it was never being rendered!