Only load FreeImage library once (instead of on-demand)
It's funny to look back on some older chunks of PD code and see how
paranoid I was about conserving memory.  As an example, FreeImage was
previously loaded manually every time it was needed, then unloaded when
done.  This introduced an unfortunate amount of overhead for all things
FreeImage-related, and it was time to fix the problem.

As of this commit, PD loads FreeImage once at program initialization,
then unloads it at program termination.  This should speed up all
functions that use the library (which is many), while also removing some
potential leaks in places where I previously forgot to free it after
errors or failures!