Kill the main form's persistent progress bar/message window
*ominous music*

(Relates to #106)

This is the first of many updates related to Issue #106, "Rework
progress bar / message area".

In this update, I have made the main form's progress bar invisible, and
instead display progress bar updates on individual child windows.  The
old custom progress bar class has also been replaced by use of a system
progress bar, and a number of other minor adjustments have been made to
the way progress is displayed and internally handled.

When an action has completed, the progress bar is hidden, meaning we no
longer waste space displaying an empty progress bar.  (Finally!)

Scroll bar existence and placement is not taken into account when
displaying the progress bar - I haven't decided if this is worth the
trouble yet, as the user cannot interact with the window during actions
anyway, but will revisit as I continue work on this.

Please note that a lot of things are deactivated by this change,
particularly batch conversion progress reports (which will need to be
completely reworked) and all user-facing messages, which are being moved
to a new location separate from the progress bar.  These features will
be restored soon.