Fix critical command bar error on compiled .exe and high-DPI screens
Man, this was unpleasant to track down.  Many thanks to RR for
discovering the connection between high-DPI screens and the error.

It seems that on a non-standard DPI (e.g. one that differs from the DPI
on which a project was originally compiled), VB performs a final resize
to user controls before they are unloaded.  This causes a problem in the
command bar, which forcibly unloads its parent control when OK/Cancel is
pressed and all related events have finished - specifically, VB attempts
to resize the user control *after* the parent has been unloaded, which
throws RTE 398 "Client site not available" because the command bar bases
its size and position off the parent's size and position - which cannot
be determined, as the parent has been unloaded!

The solution is to simply catch and ignore that error, but let me just
state that it's asinine for VB to raise additional resize events for
user controls only on DPIs that differ from the DPI used when compiling.