Add ICC profile support for CMYK JPEG, TIFF, and PSD files
Didn't think this was going to be possible, so this is a happy victory.
A bit of trickery is required to coax raw CMYK data into a 32bpp DIB,
and MS didn't help matters any by not listing KYMC order transformation
constants on MSDN, but eventually I sorted everything out!

FreeImage is currently required for the transformation.  I'm
investigating ways to get raw CMYK data out of GDI+, instead of having
it auto-converted to RGB, but so far haven't found a solution.