Color-managed CMYK conversion added to GDI+ fallbacks
Curiosity got the better of me, and I found a way to do it.  :)
Surprisingly simple!

This commit also includes optimizations for the past few CMYK-related
commits.  Basically, the primary load routine is now notified if an
external function already handled color management for an image, and it
will suspend its own handling accordingly.