Improve ExifTool shutdown behavior
(Debugging asynchronous code in VB6 is a joy, let me tell you...)

It's unpredictable, but sometimes ExifTool will not exit normally as
part of the PD shutdown process.  Explicitly giving it a bit of time to
respond to its exit instruction seems to solve the issue on my Win 7 PC,
but I'll need to test this on my other machines to be certain the fix is
universal.