Move image size, mouse coords, and messages to new status bar
Visual appearance is not final - this is just an initial commit to get
the ball rolling.  I've had to toggle off some messages to make sure
that only messages relevant to the current image are displayed, but
otherwise, it seems to work pretty well.  Persistent messages for each
image is great!