Start repair work on batch processing tool
Batch processing was mostly destroyed by the window manager overhaul and
other 6.2 interface changes.  I have now begun work on getting batch
processing back into a workable state.

The ultimate goal for batch conversion is to suspend all interface
actions while batch conversion is running (e.g. displaying images
on-screen and everything that goes with that).  This is a massive
undertaking, as pretty much everything in the program assumes that image
windows are visible, but when completed it will result in a much cleaner
and faster batch processor.

To that end, a new panel must be added to the batch convert tool.  This
panel will display running updates of a given batch process, and it will
be the only source of visual feedback while a batch conversion is
running.

This is going to be a fairly long and arduous process, so please - no
testing batch processing until all fixes are in place!  :)