Fix batch processor
Fixes #110.

Batch processor functionality *should* be restored as of this commit.
Since there is no longer a standalone message window (separate from a
loaded image), a new panel has been added to the batch wizard, and it is
used to display process.

Canceling a batch process midway has also been restored.  Instead of the
ESC key, a dedicated Cancel button is now provided.