Update checks are now asynchronous
Fixes #109.

With thanks to @Kroc for not only raising this issue in the first place,
but for providing the asynchronous download control ("bluDownload") that
made this task much easier.

Now, standard update checks (still fired every 10 days) will silently
download the latest update XML file to the user's /Data subfolder.  At
each run, PhotoDemon checks for the presence of this file, and updates
are processed accordingly.

As a convenience to the user, on-demand update checks can still be
performed from the Help -> Check for updates menu.