Add horizontal enlarge support to content-aware resize
*Significantly* faster than seam removal, as we do not need to
recalculate the seam map for every insertion.  So that's good news.

No resampling is done at present, so artifacts may be evident for large
resizes.  I'll look at revising this in the future.

Unlike the original seam carve algorithm, I have modified my version to
support any level of resize, even those 2+ times larger than the source
image.  This is done by dynamically regenerating the energy map every
time the image exceeds 150% of the original.  (Note that the 150%
threshold is effectively arbitrary; a better algorithm would allow the
user to select this value, but I don't have plans to implement that at
present.)

I have also designed this algorithm to be friendly to user-selected
preservation and removal maps, so it should work without incident when
those are added!