Menu icon and other minor improvements to content-aware resize
I've gone back and forth on what to call this function.  Originally I
used "smart resize", but have since decided against that.  "Smart
resize" implies that other resizes are *not* smart, which in turn could
make the user wonder "why I would I ever use not-smart resizing?"

Liquid rescale, as used in GIMP, etc, is not descriptive to a novice.
It's short and clever, but not intuitive.

So content-aware resize it is.  Though lengthy, this properly
distinguishes it from normal resizing as a "separate" but not
necessarily "better" tool - e.g. one to be used when appropriate!

Once I add a small paint window to the dialog, so users can select areas
to specifically maintain, the tool's purpose should be even more clear
to beginners.