Display warning after live translation change
I hate modal dialog boxes and/or "bandaid-style" fixes like this, but
the work required to live-update tooltips and other minor text is
extraordinary... so I'm calling this the lesser of two evils.

Also, language should only ever be changed once (from the default to the
user's choice) so I'm not too concerned about the workflow impact of
this dialog box.