Add "click to select coordinate" feature to preview control
A number of tools could benefit from allowing the user to select custom
coordinates for an operation (for example, changing the center point of
the vignetting effect).  The preview tool can now raise such events if
desired.

As part of a broader design decision in PD, coordinates are returned as
floating-point values in the range [0, 1], where 1 represents image
width or height.  Coordinate ratios are far more useful than absolute
pixel values if actions are to be saved as part of a macro.  (For
example, consider adding copyright text to the bottom-left corner of a
set of images.  By using coordinate ratios, even images of different
sizes will all have the text placed in the correct position!)