Change a bunch of filenames to match program updates
I hate having to do this manually as GitHub for Windows is too dumb to
note the renames itself - but oh well!

Many PD filenames have historical precedents that no longer make sense.
(Or the original filenames were just stupid, e.g. "Filters2.bas").  So
it was time to update a number of them.