Rename all "pdLayer" code as "pdDIB"
I hate mass commits like this, but sometimes its necessary.  (120 files
to be committed, ugh.)

Now that we are well on our way toward implementing layers, the time has
come to fix a rather egregious oversight on my part.  I originally
called PD's image data class "pdLayer", as I found it more intuitive to
work with the concept of layers rather than a technical term like DIB.

The problem is that actual layers (as seen in something like Photoshop)
have a ton of overhead specific to layers as a meta-object.  For
example, text attachments for text layers, style and opacity data, names
and descriptions - and all the PD code that works with raw image data
(e.g. every effect and tool!) shouldn't have to initialize a bunch of
layer metadata just to create and manipulate image bits.

So it was finally time to call pdLayer what it is - a DIB wrapper - and
change all program variables/functions/comments to match.  Obviously
this was a massive search-and-replace operation, and though I did much
of it manually just to be safe, it's possible some errors crept in.
I'll be doing a more thorough test of the program in the coming days to
be sure everything is as it should be.

(Soon I will be creating a new pdLayer class, which will actually hold
honest-to-god layer data.)