Add special handling for multiline EXIF comments
Multiline EXIF comments are technically a violation of the spec (which
explicitly states that carriage return and linefeed characters are NOT
to be used), but as with most specs, some software violates it.  PD now
checks for this special case and handles it gracefully, retrieving the
comment in full and replacing linebreaks with spaces.