Various metadata handling improvements
- Metadata export preferences are now stored on a per-image basis.  By
default, each image starts with the "use program-wide metadata storage"
setting.  Some dialogs (e.g. JPEG) can then allow the user to override
the setting for just that image.
- Binary metadata is no longer extracted, but its presence and size in
the file is marked in the metadata browser.  This way, users can see if
an image has (for example) an embedded thumbnail and its size, but
ExifTool won't waste time extracting the actual thumbnail.
- Internal ExifTool instructions in various values (e.g. "use -b option
to extract binary data") are now detected and parsed out