Improved treatment of technical JPEG metadata
Some technical JPEG data (YCbCr positioning, etc) can be stored in the
file's EXIF or XMP sections.  When re-saving a JPEG, such data may have
changed - so instead of blindly preserving it, rewrite it to match the
current technical data.

Also, when in non-privacy mode, PD is now listed as the last edit
software.  When in privacy mode, all software data is removed from the
file, and I have added code to prevent ExifTool from automatically
noting itself as the XMP Toolkit if XMP data is present.