Core Autosave framework now in place
After an unsafe shutdown, the next time PD is run, a list of Autosave
entries will be compiled.  These entries are then verified for
integrity, then presented to the user for potential restoration.

Images that were loaded but not modified will not be presented, as they
will not have any Undo/Redo data available.

Images that were loaded but were *not* marked as "already on disk", e.g.
images pasted from the clipboard or imported from a scanner, now have
special autosave data generated when they are first loaded, as a
failsafe.

Throughout the course of normal program operation, pdImages will write
XML-based "autosave" entries to a new Autosave folder (in the user's
/Data subfolder).  These XML entries contain extra information about the
image that might be helpful when reconstructing the original session.  I
don't make full use of that data at present, but theoretically I could
in the future.

There is still some work to be done; specifically, images that user
marks for restoration must actually be loaded!  This is coming soon.