First draft of new resize user control
See #115.

This first draft was all about moving the resize UI code out of the
Image Resize dialog and into a standalone control and getting everything
working again.  This is mostly done.

Integration with the command bar (e.g. last-used settings) is not
finished yet, and the control does not yet theme properly, particularly
its font settings.  I'm working on it.