Add dev warning to Resize dialog
I've got some good updates to sync, but the resize dialog will remain a
bit wonky until work on the new resize UC completes.  As I'm not doing
the work on a separate branch (shame on me, I know), I'm going to simply
add a warning to the dialog for now.