Initial bulk of viewport overhaul
Relates to #98, #117, #89, #106.

This is the first stage of the overhaul described in #98.  The goal with
this commit is to basically make PD minimally functional again, but
using a custom "canvas" user control instead of standalone windows for
displaying and interacting with images.

I haven't tried compiling this commit, but you can run the software from
the IDE and do basic image edit tasks.  Many more commits will be
incoming as I do more detailed testing...