Additional failsafe against image tabstrip corruption
I think I've finally isolated the random bug where the image tabstrip
renders thumbnails as if it's top-aligned, despite it clearly being
left-aligned.

Hard crashes can cause PD to rebuild the user preferences file from
scratch.  When this happens, it defaults to a top-aligned image
tabstrip.  However, if the image tabstrip's previous window position was
left-aligned, the window manager will use that data preferentially, so
we end up with a left-aligned tabstrip that renders thumbnails as if
it's top-aligned!

To remedy this, PD will now rewrite the window location file from
scratch if it's forced to rewrite the main preferences file.

(This should help protect against random toolbar disappearances as
well.)