Greatly improve image DPI detection
By default, FreeImage and GDI+ will now try to retrieve image resolution
from the file header (if supported).  If ExifTool is available, it will
perform a second check for resolution data inside the image's metadata,
which is actually a much more common place to find it.

Resolution is also stored as floating-point now, for precision reasons.