Add support for Percent-based Resize
Relates to #115

As stated, PD's Resize UC now supports percent-based resize
measurements.  This is actually a very large commit, as a whole
framework has been built for various measurement units, which will make
it much easier to add new units in the future.

There are still some kinks to work out with the way the UC interacts
with the actual resize code, particularly during macros.  The unit of
measurement will need to be passed to the resize sub, so something like
"resize 50%" can be used as a valid macro option.  (Rather than at
present, where "resize 50%" will store the current image's width/height
at 50% as the macro resize event.)