More bugfixes to resize UC
PPI and PPCM calculations were reversed.  (It's confusing, as typically
inch measurements are *smaller* than cm measurements, but for PPI/PPCM
the opposite is true!)

Reset values wasn't resetting resolution measurement to PPI.

Attempting to forcibly set a measurement in pixels (as used by Randomize
on the command bar) was not properly restoring the width/height units to
pixels.