New API scroll bar class
This class is an attempt to solve some long-running issues relating to
scroll bar usage in PD.  Note that this is *not* an implementation of
automatic window scroll bars, e.g. system-managed bars used to scroll
the client area of a window.  This implements a standalone scroll bar
*control*, to be used however and wherever you like, with full control
over creation, destruction, size and positioning.  (In the case of PD,
scroll bars are used as part of many user controls, e.g. sliders, text
up/down, etc, so a fine-grained level of control is necessary!)

The biggest problem with scroll bars in PD is that they lose track of
the mouse if heavy background processing occurs during the scroll event.
For example, when using a scroll bar to change blur radius, the scroll
bar will often fail to recognize the user releasing the mouse, causing
the bar to remain in a "down" state even after the mouse is released and
moves away.  This leads to a runaway state where the bar continues to
jump all over as the user moves the mouse, but because they have
released the button, there is no way to solve the issue except to click
the bar again - which may also fail to register a mouse release, causing
endless mouse capture woes!

The 2-byte limit on VB scroll bars is also annoying, especially since
our various user controls allow for floating-point control, which means
that with two significant digits, a scroll bar can only support up to
326.75 - not acceptable for many tasks.

Fixing the "lost track of mouse" proved more difficult than I
anticipated, as system scroll bars also suffer this problem.  I think I
have finally solved the issue by manually tracking mouse up events
myself - which are returned properly by the system, even if the scroll
bar tends to miss them - and using such events to forcibly disable the
scroll bar.

The class handles both horizontal and vertical bars, and works by
attaching itself to an hWnd of your choosing (typically a picture box).
It will automatically resize itself to fit the parent hWnd, so size and
position can be controlled via existing code by simply resizing the
container hWnd.  (I needed this capability as PD user controls use
larger and smaller scroll bars than the system standard, and client-area
scroll bars have their size automatically managed by Windows.)

It should be relatively bug-free, including in the IDE, but it will
certainly get a lot more testing as I convert all scroll bars in the
program to it...